chore(deps): bump rusqlite 0.37 -> 0.40.1 and adapt to FromSql change
rusqlite 0.40 removed the `usize: FromSql` impl, so the count query in
`Database::count_*` no longer compiled. Read the COUNT as `i64` and cast
to `usize`. Rebased on current main (toolchain 1.95.0 + clippy fixes) so
`cargo clippy --all-targets --all-features -- -D warnings` is clean.

+++ src/db.rs
@@ -645,11 +645,12 @@ impl<'a> Database<'a> {
         query.push_str(" AND gradingStatus != 0");
 
         let param_refs: Vec<&dyn rusqlite::ToSql> = params.iter().map(|p| p.as_ref()).collect();
-        let count: usize = self
+        // rusqlite 0.40 dropped the `usize: FromSql` impl; read as i64 and cast.
+        let count: i64 = self
             .conn
             .query_row(&query, param_refs.as_slice(), |row| row.get(0))?;
 
-        Ok(count)
+        Ok(count as usize)
     }
 
     // Transaction helpers
